Primer lectura a BDD en Entidad Jugador

+++ Dominio/Jugador.cs
@@ -18,16 +18,14 @@ namespace Dominio
         }
 
         public int IdJugador { get; set; }
-        public float Altura { get; set; }
-        public float Peso { get; set; }
+        public int Altura { get; set; }
+        public decimal Peso { get; set; }
         public string Posicion { get; set; }
         public Categoria Categoria { get; set; }
         public EstadoJugador Estado { get; set; }
         public List<Incidencia> Incidencias { get; set; }
         public List<Asistencia> Asistencias { get; set; }
         public List<Reporte> Reportes { get; set; }
-
-
 
     }
 }
